Move specific test into month module
